+++ terraform/event/event-dispatcher.tf
@@ -3,7 +3,7 @@ resource "helm_release" "event-dispatcher-service-application" {
   name       = "event-dispatcher-service-application"
   chart      = "${var.helm_repository}/event-dispatcher-service/application"
 #  version    = "1.1.2"
-  namespace  = "message-broker"
+  namespace  = "event"
   create_namespace = true
   timeout = var.helm_timeout