+++ terraform/infra/istio/label.sh
@@ -3,7 +3,7 @@ kubectl create namespace default; kubectl label namespace default "istio.io/data
 
 kubectl create namespace core; kubectl label namespace core "istio.io/dataplane-mode"="ambient"
 kubectl create namespace example; kubectl label namespace example "istio.io/dataplane-mode"="ambient"
-kubectl create namespace message-broker; kubectl label namespace message-broker "istio.io/dataplane-mode"="ambient"
+kubectl create namespace event; kubectl label namespace event "istio.io/dataplane-mode"="ambient"
 
 kubectl create namespace oidc; kubectl label namespace message-broker "istio.io/dataplane-mode"="ambient"
 kubectl create namespace monitoring; kubectl label namespace monitoring "istio.io/dataplane-mode"="ambient"